Creating a new player and game rooms
Added tables to support the game structure.  Created a new player method that will check for an available room and add the player to that game room or create a new game room and populate the questions.  Returns the game room ID and the player ID to be utilized on the front end.